ProjectShape + refactor

+++ routes/paintingBoard2.js
@@ -1,4 +1,5 @@
 const PaintingProject = require('../model/paintingProject');
+const CONST = require('../public/js/canvas/constants');
 
 const CONTROLLER = '/paintingBoard2'
 module.exports = (app) => {
@@ -50,11 +51,14 @@ module.exports = (app) => {
      * Response: { success: true }
      */
     app.delete(CONTROLLER + '/projects', async (req, res) => {
-        const id = req.query.id;
-        if (id) {
-            await PaintingProject.deleteOne({ _id: id });
+        if (req.session.passport && req.session.passport.user) {
+            const userId = req.session.passport.user._id;
+            const id = req.query.id;
+            if (id) {
+                await PaintingProject.deleteOne({ _id: id, userId }).exec();
+            }
+            res.send({ success: true });
         }
-        res.send({ success: true });
     });
 
     /**
@@ -62,11 +66,10 @@ module.exports = (app) => {
      * Response: Array of project objects
      */
     app.get(CONTROLLER + '/projects/all', async (req, res) => {
-        let user;
         if (req.session.passport && req.session.passport.user) {
-            user = req.session.passport.user;
-            const projects = await PaintingProject.find({ userId: req.session.passport.user._id }) || [];
-            res.send(projects);
+            const userId = req.session.passport.user._id;
+            const projects = await getPaintingProjectsByUserId(userId);
+            res.send(projects || []);
         } else {
             res.redirect('/');
         }
@@ -81,11 +84,16 @@ module.exports = (app) => {
         if (!id) {
             return res.status(400).send('Project id is required');
         }
-        let user;
+
         if (req.session.passport && req.session.passport.user) {
-            user = req.session.passport.user;
-            const projects = await PaintingProject.find({ userId: req.session.passport.user._id, _id: id }) || [];
-            res.send(projects[0]);
+            const userId = req.session.passport.user._id;
+            const project = await getPaintingProjectByIdAndUser(id, userId);
+            if (!project) {
+                return res.status(404).send('Project not found');
+            }
+            await addProjectShapes(project, userId);
+            res.send(project);
+
         } else {
             res.status(401).send('Unauthorized');
         }
@@ -99,22 +107,69 @@ module.exports = (app) => {
      * If id is provided, updates the existing project. Otherwise, creates a new project.
      */
     app.post(CONTROLLER + '/save', async (req, res) => {
-        const projectData = req.body.project;
-        const id = req.query.id || projectData._id;
-        let project;
-        if (id) {
-            project = await PaintingProject.findById(id).exec();
+
+        if (req.session.passport && req.session.passport.user) {
+            const userId = req.session.passport.user._id;
+            const projectData = req.body.project;
+            const id = req.query.id || projectData._id;
+            let project;
+            if (id) {
+                project = await PaintingProject.findById(id).exec();
+                if (project.userId.toString() !== userId) {
+                    return res.status(403).send('Forbidden');
+                }
+            } else {
+                res.status(400).send('Project id not found');
+            }
+            if (!project) {
+                project = new PaintingProject();
+            }
+            project.userId = userId;
+            project.name = projectData.name;
+            project.layers = projectData.layers;
+            project.canvas = projectData.canvas;
+            project.dateCreated = projectData.dateCreated;
+            project.dateModified = Date.now();
+            const answer = await project.save();
+            res.send({ id: answer._id });
+        } else {
+            res.status(401).send('Unauthorized');
         }
-        if (!project) {
-            project = new PaintingProject();
-        }
-        project.userId = req.session.passport.user._id;
-        project.name = projectData.name;
-        project.layers = projectData.layers;
-        project.canvas = projectData.canvas;
-        project.dateCreated = projectData.dateCreated;
-        project.dateModified = Date.now();
-        const answer = await project.save();
-        res.send({ id: answer._id });
     });
+
+    async function addProjectShapes(project, userId) {
+        if (!project.layers || project.layers.length === 0) {
+            return;
+        }
+        const projects = {}
+        for (const layer of project.layers) {
+            for (const shape of (layer.shapes || [])) {
+                if (shape.desc === CONST.PROJECT_SHAPE) {
+                    const projectId = shape.projectId;
+                    if (projects[projectId]) {
+                        shape.layers = projects[projectId].layers;
+                        continue;
+                    } else {
+                        const project = await getPaintingProjectByIdAndUser(projectId, userId);
+                        if (project) {
+                            projects[projectId] = project;
+                            shape.layers = project.layers;
+                            shape.jajaja = "jajaja"
+                        }
+                    }
+                }
+            }
+        }
+        return project;
+    }
+
+    async function getPaintingProjectByIdAndUser(id, userId) {
+        const project = await PaintingProject.findOne({ _id: id, userId }).exec();
+        return addProjectShapes(project.toObject(), userId);
+    }
+
+    async function getPaintingProjectsByUserId(userId) {
+        const projects = await PaintingProject.find({ userId }).exec();
+        return Promise.all(projects.map(project => addProjectShapes(project.toObject(), userId)));
+    }
 }
